backend/requirements.txt
fastapi==0.115.6
pillow==11.0.0
torch==2.5.1
transformers==4.47.0
uvicorn==0.34.0